+++ hutool-core/src/main/java/cn/hutool/core/builder/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * 建造者工具<br>
+ * 用于建造特定对象或结果
+ * 
+ * @author looly
+ *
+ */
+package cn.hutool.core.builder;