+++ hutool-extra/src/main/java/cn/hutool/extra/template/engine/velocity/StringResourceLoader.java
@@ -1,6 +1,5 @@
 package cn.hutool.extra.template.engine.velocity;
 
-import java.io.ByteArrayInputStream;
 import java.io.InputStream;
 
 import org.apache.commons.collections.ExtendedProperties;
@@ -8,18 +7,18 @@ import org.apache.velocity.exception.ResourceNotFoundException;
 import org.apache.velocity.runtime.resource.Resource;
 import org.apache.velocity.runtime.resource.loader.ResourceLoader;
 
+import cn.hutool.core.io.IoUtil;
 import cn.hutool.core.util.CharsetUtil;
 
 public class StringResourceLoader extends ResourceLoader{
 
 	@Override
 	public void init(ExtendedProperties configuration) {
-		
 	}
-
+	
 	@Override
 	public InputStream getResourceStream(String source) throws ResourceNotFoundException {
-		return new ByteArrayInputStream(source.getBytes(CharsetUtil.CHARSET_UTF_8));
+		return IoUtil.toStream(source, CharsetUtil.CHARSET_UTF_8);
 	}
 
 	@Override