+++ hutool-extra/src/test/java/cn/hutool/extra/template/TemplateUtilTest.java
@@ -1,6 +1,11 @@
 package cn.hutool.extra.template;
 
+import java.io.File;
+import java.util.HashMap;
+import java.util.Map;
+
 import org.junit.Assert;
+import org.junit.Ignore;
 import org.junit.Test;
 
 import cn.hutool.core.lang.Dict;
@@ -22,13 +27,13 @@ public class TemplateUtilTest {
 
 	@Test
 	public void createEngineTest() {
-		//默认模板引擎，此处为Beetl
+		// 默认模板引擎，此处为Beetl
 		Engine engine = TemplateUtil.createEngine(new TemplateConfig());
 		Template template = engine.getTemplate("hello,${name}");
 		String result = template.render(Dict.create().set("name", "hutool"));
 		Assert.assertEquals("hello,hutool", result);
 	}
-	
+
 	@Test
 	public void beetlEngineTest() {
 		// 字符串模板
@@ -43,7 +48,7 @@ public class TemplateUtilTest {
 		String result2 = template2.render(Dict.create().set("name", "hutool"));
 		Assert.assertEquals("hello,hutool", result2);
 	}
-	
+
 	@Test
 	public void rythmEngineTest() {
 		// 字符串模板
@@ -57,7 +62,7 @@ public class TemplateUtilTest {
 		String result2 = template2.render(Dict.create().set("name", "hutool"));
 		Assert.assertEquals("hello,hutool", result2);
 	}
-	
+
 	@Test
 	public void freemarkerEngineTest() {
 		Engine engine = new FreemarkerEngine(new TemplateConfig("templates", ResourceMode.CLASSPATH));
@@ -71,9 +76,14 @@ public class TemplateUtilTest {
 		Engine engine = new VelocityEngine(new TemplateConfig("templates", ResourceMode.CLASSPATH));
 		Template template = engine.getTemplate("templates/velocity_test.vtl");
 		String result = template.render(Dict.create().set("name", "hutool"));
-		Assert.assertEquals("hello,hutool", result);
+		Assert.assertEquals("你好,hutool", result);
+		
+		engine = new VelocityEngine(new TemplateConfig("templates", ResourceMode.STRING));
+		template = engine.getTemplate("你好,$name");
+		result = template.render(Dict.create().set("name", "hutool"));
+		Assert.assertEquals("你好,hutool", result);
 	}
-	
+
 	@Test
 	public void enjoyEngineTest() {
 		// 字符串模板
@@ -82,7 +92,7 @@ public class TemplateUtilTest {
 		String result = template.render(Dict.create().set("x", 1));
 		Assert.assertEquals("124", result);
 	}
-	
+
 	@Test
 	public void thymeleafEngineTest() {
 		// 字符串模板
@@ -90,5 +100,17 @@ public class TemplateUtilTest {
 		Template template = engine.getTemplate("<h3 th:text=\"${message}\"></h3>");
 		String result = template.render(Dict.create().set("message", "Hutool"));
 		Assert.assertEquals("<h3>Hutool</h3>", result);
+	}
+	
+	@Test
+	@Ignore
+	public void renderToFileTest() {
+		Engine engine = new BeetlEngine(new TemplateConfig("templates", ResourceMode.CLASSPATH));
+		Template template = engine.getTemplate("freemarker_test.ftl");
+
+		final Map<String, Object> bindingMap = new HashMap<>();
+		bindingMap.put("name", "aa");
+		File outputFile = new File("e:/test.txt");
+		template.render(bindingMap, outputFile);
 	}
 }