+++ hutool-http/src/main/java/cn/hutool/http/useragent/Platform.java
@@ -1,5 +1,6 @@
 package cn.hutool.http.useragent;
 
+import java.util.ArrayList;
 import java.util.List;
 
 import cn.hutool.core.collection.CollUtil;
@@ -19,13 +20,25 @@ public class Platform extends UserAgentInfo {
 	 * 支持的移动平台类型
 	 */
 	public static final List<Platform> mobilePlatforms = CollUtil.newArrayList(//
+			new Platform("Windows Phone", "windows (ce|phone|mobile)( os)?"), //
 			new Platform("iPad", "ipad"), //
 			new Platform("iPod", "ipod"), //
 			new Platform("iPhone", "iphone"), //
 			new Platform("Android", "android"), //
-			new Platform("Windows Phone", "windows (ce|phone|mobile)( os)?"), //
 			new Platform("Symbian", "symbian(os)?"), //
 			new Platform("Blackberry", "blackberry") //
+	);
+
+	/**
+	 * 支持的桌面平台类型
+	 */
+	public static final List<Platform> desktopPlatforms=CollUtil.newArrayList(//
+			new Platform("Windows", "windows"), //
+			new Platform("Mac", "(macintosh|darwin)"), //
+			new Platform("Linux", "linux"), //
+			new Platform("Wii", "wii"), //
+			new Platform("Playstation", "playstation"), //
+			new Platform("Java", "java") //
 	);
 	
 	/**
@@ -33,16 +46,9 @@ public class Platform extends UserAgentInfo {
 	 */
 	public static final List<Platform> platforms;
 	static {
-		platforms = CollUtil.newArrayList(//
-				// 移动平台
-				new Platform("Windows", "windows"), //
-				new Platform("Mac", "(macintosh|darwin)"), //
-				new Platform("Linux", "linux"), //
-				new Platform("Wii", "wii"), //
-				new Platform("Playstation", "playstation"), //
-				new Platform("Java", "java") //
-		);
+		platforms=new ArrayList<Platform>(13);
 		platforms.addAll(mobilePlatforms);
+		platforms.addAll(desktopPlatforms);
 	}
 
 	/**